LayoutTests/http/wpt/webauthn/public-key-credential-get-failure-local-silent.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with no matched credentials in a mock local authenticator.
PASS PublicKeyCredential's [[get]] with no matched credentials in a mock local authenticator. 2nd
PASS PublicKeyCredential's [[get]] without user consent in a mock local authenticator.
